+++ LootEnums.cs
@@ -1,0 +1,11 @@
+namespace LootManager
+{
+    public enum EquipmentTierSetting
+    {
+        All,
+        NormalOnly,
+        BlessedOnly,
+        GodlyOnly,
+        BlessedAndUp
+    }
+}